pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gemini-chat"
version = "1.0.0"
description = "基于 Google Gemini API 与 PySide6 的跨平台 AI 聊天客户端"
authors = [
    {name = "Gemini Chat Team", email = "team@geminichat.dev"}
]
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
keywords = ["ai", "chat", "gemini", "pyside6", "desktop"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "PySide6>=6.5.0",
    "google-genai>=1.0.0",  # 使用最新的Google GenAI SDK
    "python-dotenv>=1.0.0",
    "toml>=0.10.2",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-qt>=4.2.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
build = [
    "pyinstaller>=5.0.0",
]

[project.urls]
Homepage = "https://github.com/gemini-chat-team/gemini-chat"
Repository = "https://github.com/gemini-chat-team/gemini-chat.git"
Issues = "https://github.com/gemini-chat-team/gemini-chat/issues"

[project.scripts]
gemini-chat = "geminichat.main:main"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["geminichat*"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true


